+++ src/main/java/com/azer467/saphyre/events/world/SaphyreWorldEvents.java
@@ -1,8 +1,8 @@
-package com.azer467.saphyre.core.world.event;
+package com.azer467.saphyre.events.world;
 
 
 import com.azer467.saphyre.SaphyreMetadata;
-import com.azer467.saphyre.core.world.gen.SaphyreOreGeneration;
+import com.azer467.saphyre.world.SaphyreOreGeneration;
 import net.minecraftforge.event.world.BiomeLoadingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;