+++ CHANGELOG.md
@@ -13,10 +13,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - Advancements
+- Saphyre Entity
 
 ### Changed
 
 - Code Cleanup
+- Folder Structure
 
 ## [0.3.0-118-FORGE] - 06/06/2022